Merge introduction into TYPO3 Explained
Newer versions of this manual were already merged into TSref /
TYPO3 Explained (since 9.5).

We have now merged the major parts of this manual into
"TYPO3 Explained" and the
the "TypoScript reference" (in line with the newer branches).

Links to redirect the user have been added to the pages and
the content removed.